file restructure and adding booking main page
